Added UserDao changing password and roles functionalities Added UserDao and Roles services, interfaces, repos and controllers

+++ src/main/java/com/carecure/medsysten/security/models/UserRepository.java
@@ -1,6 +1,6 @@
 package com.carecure.medsysten.security.models;
 
 import org.springframework.data.repository.CrudRepository;
-public interface UserRepository extends CrudRepository<UserDao, Integer> {
+public interface UserRepository extends CrudRepository<UserDao, Long> {
     UserDao findByUsername(String username);
 }
